requirements.txt
Flask==3.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
docx2python==2.0.2
gunicorn==21.2.0
pytest==7.4.3
pytest-mock==3.12.0